move pawn test to its own file and rename it back to PawnTest
