+++ tests.py
@@ -1,4 +1,5 @@
 from tests.board_test import BoardTests
+from tests.pawn_test import PawnTest
 from tests.piece_test import *
 import unittest
